move mapEntry to utils
