+++ src/api/agent.js
@@ -1,60 +1,11 @@
 import * as contentful from 'contentful';
 import _ from 'lodash';
 
-export const mapAgentEntryToEntitiy = (entry) => {
-  return {
-    id: _.get(entry, 'sys.id'),
-    name: _.get(entry, 'fields.name'),
-    lastname: _.get(entry, 'fields.lastname'),
-    image: _.get(entry, 'fields.image.fields.file.url'),
-    rate: {
-      rating: _.get(entry, 'fields.rating'),
-      count: 10,
-    },
-    phone: _.get(entry, 'fields.phone'),
-    company: _.get(entry, 'fields.company'),
-    specialization: _.get(entry, 'fields.specialization'),
-    licenseNumber: _.get(entry, 'fields.licenseNumber'),
-    about: _.get(entry, 'fields.about'),
-    UID: _.get(entry, 'fields.UID'),
-    area: _.get(entry, 'fields.area'),
-  };
-};
-
-const mapAgentReivewEntryToEntitiy = (entry) => {
-  return {
-    id: _.get(entry, 'sys.id'),
-    createdAt: _.get(entry, 'sys.createdAt'),
-    updatedAt: _.get(entry, 'sys.updatedAt'),
-    ...entry.fields,
-  };
-};
-
-const mapPropertyEntryToEntitiy = (entry) => {
-  return {
-    id: _.get(entry, 'sys.id'),
-    createdAt: _.get(entry, 'sys.createdAt'),
-    updatedAt: _.get(entry, 'sys.updatedAt'),
-    ...entry.fields,
-    mainImage: _.get(entry, 'fields.mainImage.fields.file.url'),
-    room: {
-      bedroom: _.get(entry, 'fields.bedroom', 0),
-      bathroom: _.get(entry, 'fields.bathroom', 0),
-    },
-  };
-};
-
-const mapActivityEntryToEntitiy = (entry, includes) => {
-  // Add linkEntry to by
-  const agentEntry = _.find(includes.Entry, includeEntry => includeEntry.sys.id === entry.fields.by.sys.id);
-  return {
-    id: _.get(entry, 'sys.id'),
-    createdAt: _.get(entry, 'sys.createdAt'),
-    updatedAt: _.get(entry, 'sys.updatedAt'),
-    ...entry.fields,
-    by: _.get(agentEntry, 'fields.name', ''),
-  };
-};
+// Content Mapper
+import mapAgentEntryToEntity from './utils/mapAgentEntryToEntity';
+import mapActivityEntryToEntity from './utils/mapActivityEntryToEntity';
+import mapReferenceEntryToEntity from './utils/mapReferenceEntryToEntity';
+import mapPropertyEntryToEntity from './utils/mapPropertyEntryToEntity';
 
 const client = contentful.createClient({
   space: process.env.REACT_APP_SPACE,
@@ -73,14 +24,14 @@ export const getAgentEntries = ({ text, area }) => {
   return client.getEntries(cleanQuery).then((response) => {
     return {
       ...response,
-      items: _.map(response.items, item => mapAgentEntryToEntitiy(item)),
+      items: _.map(response.items, item => mapAgentEntryToEntity(item)),
     };
   });
 };
 
 export const getAgentEntry = (agentId) => {
   return client.getEntries({ 'sys.id': agentId }).then((response) => {
-    return mapAgentEntryToEntitiy(_.head(response.items));
+    return mapAgentEntryToEntity(_.head(response.items));
   });
 };
 
@@ -92,7 +43,7 @@ export const getAgentReferences = (agentId) => {
   .then((response) => {
     return {
       ...response,
-      items: _.map(response.items, item => mapAgentReivewEntryToEntitiy(item)),
+      items: _.map(response.items, item => mapReferenceEntryToEntity(item)),
     };
   });
 };
@@ -105,7 +56,7 @@ export const getAgentProperties = (agentId) => {
   .then((response) => {
     return {
       ...response,
-      items: _.map(response.items, item => mapPropertyEntryToEntitiy(item)),
+      items: _.map(response.items, item => mapPropertyEntryToEntity(item)),
     };
   });
 };
@@ -121,7 +72,7 @@ export const getAgentActivities = (agentId) => {
     return {
       ...response,
       items: _.map(response.items, (item) => {
-        return mapActivityEntryToEntitiy(item, includes);
+        return mapActivityEntryToEntity(item, includes);
       }),
     };
   });
